{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "{{ cookiecutter.version }}"
readme = "readme.md"
package-mode = false

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed",
]

packages = [
    { include = "{{ cookiecutter.package_name }}", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pyview-web = "^0.0.7a0"

[tool.poetry.dev-dependencies]
pytest = "*"
ruff = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
